chore: test fix exec.node.js tests
It was not using the dirty-chai so the expect.to.not.exist was failing. It was also trying to trap SIG* with bin/sh which does not recognize those signals (dropping the SIG prefix works in sh).

This SIG* trapping was what I think was making CI tests fail so I enabled testing in CI again.