fix: fall back to path
When developing features on `ipfs`/`jsipfs` concurrently you usually
have your dev versions installed on the path.

To alleviate the tedium of having to type:

```console
$ IPFS_JS_EXEC=`which jsipfs` IPFS_GO_EXEC=`which ipfs` npm test
```

...we could just look on the path for an `ipfs`/`jsipfs` to run.

The PR uses the [`which`](https://www.npmjs.com/package/which) module
to attempt to find `ipfs`/`jsipfs` on the path if all other methods
fail.