fix: add namesys pubsub option (#293)
For implementing the tests for `IPNS over Pubsub` on `js-ipfs` using a `type: proc` daemon, we need to have the `--enable-namesys-pubsub` option properly handled. 

Moreover, I cannot add a test for it, similar to the one already available for `--enable-pubsub-experiment`, as it is not available in the last released version of `js-ipfs`. Once this is merged, I will create an Issue for adding those tests and when `IPNS over Pubsub` PR is merged and released, I will get back here for adding the proper test.