refactor: remove path and ref from module args also findBin
Since we're not passing modules to `.spawn()` there's no danger of
passing module refs over http, so we can remove the `.path` and
`.ref` arguments when specifying which `ipfs-http-client` and `ipfs`
modules to use, instead setting them up when you configure the
daemon factory, either in-proc or as a server in an `.aegir` file.

Also removes the `findBin` command magic in favour of being explicit
about which binary you want the factory to use.

If you need to support env vars like `IPFS_GO_EXEC`, do that when
you set up the factory.

BREAKING CHANGES:

- `.path` and `.ref` args removed from `ipfsModule` and `ipfsHttpModule`
- `findBin` function removed