chore: remove peer deps
We added ipfs, the http client etc as peer deps to signal to the user
that they should add the modules they depend on as deps of their project.

Starting with npm7 all peer deps get installed automatically which
defeats the purpose of our use of peer deps, so let's remove them.